requirements.txt
discord.py
asyncio
google
googletrans
websockets
urllib3
requests
